Move registration and listing of registered resources to dataset endpoint
